Change commission percentage data type to int

+++ Facturacion/DTOs/CreateSellerDto.cs
@@ -4,7 +4,7 @@
   {
     public required string FirstName { get; set; }
     public required string LastName { get; set; }
-    public byte? CommissionPercentage { get; set; }
+    public int? CommissionPercentage { get; set; }
     public bool? IsActive { get; set; }
   }
 }
